+++ inc/class.fields-controller.php
@@ -36,6 +36,7 @@ class Fields_Controller {
     }
     // add hooks for generated multilingual fields (type of those will be 'group')
     add_filter("acf/format_value/type=group", [$this, 'format_multilingual_value'], 12, 3);
+    add_filter("acf/update_value/type=group", [$this, 'before_update_multilingual_value'], 9, 4);
     add_filter("acf/update_value/type=group", [$this, 'after_update_multilingual_value'], 12, 4);
     add_action("acf/render_field/type=group", [$this, 'render_multilingual_field'], 5);
     add_filter("acf/load_value/type=group", [$this, 'inject_previous_monolingual_value'], 10, 3);
@@ -108,9 +109,13 @@ class Fields_Controller {
         'acfml_multilingual_subfield' => 1,
         'wrapper' => $wrapper,
       ]);
+      if( !empty($field['prepend']) ) {
+        $sub_field['prepend'] = acfml()->convert_urls_in_string($field['prepend'], $lang);
+      }
       // add the subfield
       $sub_fields[] = $sub_field;
     }
+    
     // Add 'required'-indicator to the groups label, if it is set to required
     $label = $field['label'];
     if( $field['required'] ) $label .= " <span class=\"acf-required\">*</span>";
@@ -168,6 +173,24 @@ class Fields_Controller {
     if( !$this->is_acfml_group($field) ) return $value;
     $language = acfml()->get_current_language();
     $value = !empty($value[$language]) ? $value[$language] : ($value[acfml()->get_default_language()] ?? null);
+    return $value;
+  }
+
+
+  /**
+   * Applies custom "acfml_sanitize_callback" to field values before saving to the database.
+   * Used for slugs
+   *
+   * @param mixed $value
+   * @param int $post_id
+   * @param array $field
+   * @return mixed
+   */
+  public function before_update_multilingual_value( $value, $post_id, $field, $value_before ) {
+    if( !$this->is_acfml_group($field) ) return $value;
+    if( is_array($value) && !empty($field['acfml_sanitize_callback']) && function_exists($field['acfml_sanitize_callback']) ) {
+      $value = array_map($field['acfml_sanitize_callback'], $value);
+    }
     return $value;
   }